--- tests/Unit/Expressions/ForClasses/DummyClasses/DwarfCavendishBanana.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Arkitect\Tests\Unit\Expressions\ForClasses\DummyClasses;
-
-final class DwarfCavendishBanana extends CavendishBanana
-{
-}
